feat: Controller 테스트 성공, 실패 분리하기 (#102)
