+++ src/test/java/com/rudkids/rudkids/interfaces/magazine/MagazineControllerTest.java
@@ -95,43 +95,6 @@ class MagazineControllerTest extends ControllerTest {
             .andExpect(status().isOk());
     }
 
-    @DisplayName("존재하지 않는 매거진을 수정할 경우 상태코드 404를 반환한다.")
-    @Test
-    void 존재하지_않는_매거진을_수정할_경우_상태코드_404를_반환한다() throws Exception {
-        doThrow(new MagazineNotFoundException())
-            .when(magazineService)
-            .update(any(), any(), any());
-
-        mockMvc.perform(put(MAGAZINE_DEFAULT_URL + "/{id}", MAGAZINE_ID)
-                .header(AUTHORIZATION_HEADER_NAME, AUTHORIZATION_HEADER_VALUE)
-                .accept(MediaType.APPLICATION_JSON)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(MAGAZINE_수정_요청())))
-            .andDo(print())
-            .andDo(document("magazine/update/failByNotFoundError",
-                preprocessRequest(prettyPrint()),
-                preprocessResponse(prettyPrint()),
-                requestHeaders(
-                    headerWithName("Authorization")
-                        .description("JWT Access Token")
-                ),
-                pathParameters(
-                    parameterWithName("id")
-                        .description("존재하지 않는 매거진 id")
-                ),
-                requestFields(
-                    fieldWithPath("title")
-                        .type(JsonFieldType.STRING)
-                        .description("새로운 제목"),
-
-                    fieldWithPath("content")
-                        .type(JsonFieldType.STRING)
-                        .description("새로운 내용")
-                )
-            ))
-            .andExpect(status().isNotFound());
-    }
-
     @DisplayName("매거진을 삭제한다.")
     @Test
     void 매거진을_삭제한다() throws Exception {
@@ -155,31 +118,6 @@ class MagazineControllerTest extends ControllerTest {
                 )
             ))
             .andExpect(status().isOk());
-    }
-
-    @DisplayName("존재하지 않는 매거진을 삭제할 경우 상태코드 404를 반환한다.")
-    @Test
-    void 존재하지_않는_매거진을_삭제할_경우_상태코드_404를_반환한다() throws Exception {
-        doThrow(new MagazineNotFoundException())
-            .when(magazineService)
-            .delete(any(), any());
-
-        mockMvc.perform(delete(MAGAZINE_DEFAULT_URL + "/{id}", MAGAZINE_ID)
-                .header(AUTHORIZATION_HEADER_NAME, AUTHORIZATION_HEADER_VALUE))
-            .andDo(print())
-            .andDo(document("magazine/delete/failByNotFoundError",
-                preprocessRequest(prettyPrint()),
-                preprocessResponse(prettyPrint()),
-                requestHeaders(
-                    headerWithName("Authorization")
-                        .description("JWT Access Token")
-                ),
-                pathParameters(
-                    parameterWithName("id")
-                        .description("존재하지 않는 매거진 id")
-                )
-            ))
-            .andExpect(status().isNotFound());
     }
 
     @DisplayName("매거진을 전체조회한다.")
@@ -234,25 +172,5 @@ class MagazineControllerTest extends ControllerTest {
                 )
             ))
             .andExpect(status().isOk());
-    }
-
-    @DisplayName("존재하지 않는 매거진을 상세조회할 경우 상태코드 404를 반환한다.")
-    @Test
-    void 존재하지_않는_매거진을_상세조회할_경우_상태코드_404를_반환한다() throws Exception {
-        doThrow(new MagazineNotFoundException())
-            .when(magazineService)
-            .find(any());
-
-        mockMvc.perform(get(MAGAZINE_DEFAULT_URL + "/{id}", MAGAZINE_ID))
-            .andDo(print())
-            .andDo(document("magazine/find/failByNotFoundError",
-                preprocessRequest(prettyPrint()),
-                preprocessResponse(prettyPrint()),
-                pathParameters(
-                    parameterWithName("id")
-                        .description("매거진 id")
-                )
-            ))
-            .andExpect(status().isNotFound());
     }
 }
